requirements.txt
Django==3.1.7
djangorestframework==3.12.2
backoff==1.10.0
celery==5.0.5
django-ratelimit==3.0.1